Fix misc. ET/reflection issues.
- Fix bow firing animation on ETs and reflections.
- Split evil twin spell code into two extra parts - SpawnEvilTwin()
which handles the ET creation, and PlaceEvilTwin() (in the EvilTwin
class) which handles ET placement.
- Casting ET on a monster will now create an instance of the EvilTwin
class, rather than using a modified monster.
- Phase now calls PhaseMinions() and UnphaseMinions() which handle what
happens to minions (including attacking ETs) when a player phases.
Reflections are still deleted, but attacking ETs will also 'phase'.
- ETs disappear when a user logs off but will now also reappear when the
user logs back on (provided the caster is still logged on and hasn't
died).
- Fixed capitalisation on ET/reflection death messages.
- Fixed monster ET animations.