Improve flagpole troop generation.
Soldier generation timer now only starts when a user enters the room. If
no soldiers are present when a user enters, one will be spawned
immediately and the timer started. If any are present, the timer starts
if not present (but no extra soldier is immediately spawned). If a troop
gen timer fires with no user present, no soldier is spawned and the
timer is not set again.

The net result of this is that at least one soldier is always present
when a user enters the room, but users cannot switch between rooms to
force extra generation (if we want more soldiers, we can decrease the
wait on the timer). This also removes unnecessary timers and prevents
making soldiers that no player ever sees (currently soldiers are always
spawning/being deleted in every MonsterRoom with a non-neutral
flagpole).