Soldier shield settings & fixes
* New setting to restrict soldier shield specialized bonuses purely to
direct attacks and spells on other soldiers. This should prevent
scenario overlap that has proven to be a problem.

* Fixed soldier shield bug that probably always prevented tracking who
logged another player in soldier fights. It always set
poLastEnemyToAttack equal to you, as if you attacked yourself, instead
of the enemy having attacked you. Also made it depend on damage actually
being done (like it was supposed to based on passed parameters).
Phasing/penning a soldier will give you a rank boost.

* Fixed up IsEnemyAttack calls, which were frankly a mess. Now it is
(and must be) always called on the attacker's soldier shield. Who is the
defender, What is the attacker.

* Added lTargets passed down to a variety of places where it may be
useful in the future (and required for this change). It will let us
discriminate on bonuses based on what is being attacked or cast on.

* Removed extraneous #damage=TRUE in FindUsing a soldier shield in
assessdamage. FindUsing doesn't do anything with that parameter.