Fix VS 2015 clientd3d compiling error.
VS 2015 has a bug preventing clientd3d\download.c from compiling due to
incorrect compiler warnings in shlobj.h. MS is aware of the issue and it
will supposedly be fixed but until then the warning has to be ignored
when including this header file.

Removed an unused constant from heat.kod.