Don't reuse room data for duplicate rooms.
- Room 'blockers' are tied to the roomdata so duplicate rooms can no
longer reuse roomdata. Creating a duplicate room will now load a copy of
that room from disk. We should likely improve this later to keep a
master copy of each available room in memory and load from this for
speed. Deleting a room will now free the associated memory.
- Added a "show blockers" admin command that takes a room's roomdata as
parameter and shows the blockers in that room.
- Fixed the missing background issue on room creation.
- Cleaned up room's Constructed() message a bit.