Handle Qormas persisting over Recreate.
Currently Qormas won't persist correctly over a recreate, as the
active event list in GameEventEngine is set to $ and the recreated
rooms won't have the Qormas decorations.

GameEventEngine will now call Recreate on any active events. When
recreated, Qormas will end itself and post a StartQormas, which is
called after rooms get recreated (as both happen in the first part
of RecreateAll). To avoid calling messages on deleted rooms, StartSnow
and EndSnow now use System's room list to determine which rooms to
affect.

Qormas will now try to delete its ending callback when EndEvent is
called, in case it is ended manually. This is posted to avoid any
potential issue (though this is already handled, post is still safer).

Added the ability to Post to object constants to facilitate the above
Posting of callback deletion.

Added a DeleteCallback message to RealTime which takes an object and
tries to delete all callbacks for that object.

Changed some Nth(i,1) calls to First(i).