Kill caster's ETs in any safe room, fix attacks on minions.
Casters will lose their ETs on entering any safe room (the previous fix
enabled this only for Inns/adv halls and not shops etc.).

Attacks on minions now use the correct room (minion's room) and not the
master/caster's room as in the case of ET this breaks down, specifically
for guilded vs. non-guilded screens given the above change to ET
deletion.