Fix room deletion bug, add rooms table printing cmd.
If the room being deleted wasn't first in the list, the previous rooms
in the list occupying that position in the rooms table were no longer
being referenced (only rooms past the deleted one). Specifically,
deleting rented rooms after a recreate dropped references to the other
room occupying that position in the table, which due to the order of
room loading tended to be around the kc1-kd4 area. This case is now
handled correctly.

Added an admin "show roomtable" command which prints the rooms table to
admin log, and allowed this issue to be debugged.