Play a sound when you receive a tell
This plays a pre-existing wav (that is not used in the game as far as I
know) whenever the user receives a tell. This is a much needed feature
for those who are AFK, looking at other windows, and so on.

Definitely open to other sounds. This one is the most appropriate I
could find.