Hold, Blind, and Dazzle visual effects
This pull request includes animated projectile graphics for Hold,
Blind, and Dazzle. It also includes an animated overlay for players to
visually display the binding rings of purple energy from being Held.

Players who are Held, Blind, or Dazzled will also glow purple, green, or
white respectively, letting other players see that they are afflicted -
and when the affliction wears off.

Related code changes:
Pass wielder from hold proc and blind proc to CastSpell.
This enables the owner of a proc'ing weapon to still shoot the proper projectile.

Added Blind, Hold, and Dazzle glows to monsters.
This commit expands the Monster class IsEnchanted function to match the
one Player has, allowing byClass searches.

It also adds a SendLightingInformation to monsters, which they did not
have at all before, generally.

Hold, Blind, and Dazzle now tell a monster's Room that the monster
changed both on cast and on expiration, so that the room shows the glow
properly.