Refactoring of `creaweap.kod`
This commit refactors the summoning of a new weapon into a single function, cleans up several syntax inconsistencies, and most significantly, deletes weapons that cannot be carried rather than dropping them on the floor. It corrects an existing bug in the class wherein which an item attribute is added to weapons that vanish.