Reblance: Wedding Ring (#194)
The wedding-ring is more a roleplayitem and was rarely used by the players, cause nobody know the defense boon before the game went open source. The item is simple overpowered on casters, with high agility, robe and cess-faction. Even a Purefighter will allways miss the caster or even a hybrid-toon (even with half purecaster stats) over the half of the time.