Remove lever area from monster check (#403)
This commit closes #402 by reducing the range of the final room's monster check by 1 grid row. Currently, the logic reaches into row 27, which is the grid square the right lever sits in. It should not be included and since it is, a monster present in that square will block the final room from finishing.