Update second wind enchantment (#389)
* Add second wind icon and message

org coded by skittles1

Add second wind icon and message

originally coded by skittles1

* Bug Fix for Icon disappearing thru system safe

skittles1 commented 5 hours ago
---
There was a hotfix to this one that resolved an issue with the icon disappearing on system save:

if ptSecondWind <> $
{
   Send(self,@ShowAddEnchantment,#what=Send(SYS,@FindSkillByNum,
         #num=SKID_SECOND_WIND),#type=ENCHANTMENT_PLAYER);
}

This needs to be added to SendEnchantmentIcons() in player.kod.

* Remove logon check that results in visual bug

This commit removes a logon check that results in a visual bug - the presence of a second Second Wind enchantment icon when logging off and then back on while enchanted with Second Wind.

* Refactor second wind to mimic personal enchantment

These commits refactor as much of second wind out of `player.kod` as possible, and updates `secwind.kod` to mimic the personal enchantment spell API, so that Player can treat it similarly.

Co-authored-by: Nagath <nagath@freenet.de>