Update mark of dishonor to allow for monster targets (#393)
* Mark of Dishonor can be casted on monsters

A level four spell that cannot be improved by casting it on monsters is
a bit to hard, it can now be casted on monsters.
If casted it on a monster who aggro a player the monster will stop its
aggro towards that player.

* Resolved merged conflict

* Disallows casting on good players, sends message

This commit disallows casting on karmically good players.

Co-authored-by: broodhill <broodhill@gmail.com>