Add rentable rooms to Ko'catan Inn (#392)
* Rented rooms for Ko'catan

This commit adds rentable rooms to Ko'catan in the same style as the
Barloque and Marion rooms. It also adds island textures that default
match the island style. Those textures are also available as choices in
Barloque and Marion.

This commit also fixes a rare initialization issue with rented rooms,
where if the default bed position was equal to 1, it would not be lifted
or textured.

This commit also fixes a debug error in statistics - rented rooms are
not in the list of rooms to have their 'number of times entered'
counted, so there's no need to check RIDs over 10000. This should
prevent a debug message happening every single time someone enters a
rented room.

* Added Ko'catan wall texture

* Fixes typos

* Removes redundant string checks

* Changed texture name

* Corrected texture scaling

* Removed unused door exit code

I felt bad doing this because it removes the unique names someone came up with for these rooms.

* Reintroduces room names to locked doors

This commits reintroduces the Inn's unique room names into the locked door message.

Co-authored-by: GarAdminTools <meridian59gar3@gmail.com>