Removes handling of animate as a config option (#452)
* Removes animate config option

This commit completely removes the handling of an animate flag in the config. As a result, animations can not be disabled and animations are not slowed when the game window is backgrounded.

* Removes remaining calls to `config.animate`

* Cleans up remaining animate references in Config

* Removes `Animate` from config

* Removes now unused `AnimationSleep()`

* Lowers animation interval for background animations

The intention behind this commit is to lower the background animation interval to be consistent with what it would be in the foreground. 40ms was a number I pulled from `firewand.kod`:189.

* Sets ANIMATE_INTERVAL to 8ms (120fps)