Refactors list building to match order
This commits refactors the building of `plShrine_Allegiances` to be consistent with the order of `plShrines`, used later in resetting the shrine allegiances during recreate.

If the two lists are not the same length for whatever reason, recreate will note this in the debug log and continue with fresh shrines.