Adds support for recreating shrines (#449)
* Support for shrines during recreate

This commits adds support for shrine allegiance and power levels to be re-populated during recreate.

* Comment updates

* Defaults `SetAllegiance` to 0 or no allegiance

* Refactors list building to match order

This commits refactors the building of `plShrine_Allegiances` to be consistent with the order of `plShrines`, used later in resetting the shrine allegiances during recreate.

If the two lists are not the same length for whatever reason, recreate will note this in the debug log and continue with fresh shrines.

* Changes brace placement for consistency