Fixes persistent effects of Umbrella (#455)
* Create .clang-format

* Revert "Create .clang-format"

This reverts commit 9019efef4b4e64942de7f8e3b7680360fae16bd2.

* Fixes persistent umbrella effects

This commit implements `RemoveEnchantmentEffects` to handle ending Umbrella outside of its timer. When a player is purged or a DM removes their enchantments, the resistance effects of Umbrella will remain. This commit ensures that in these scenarios, the resistance modifiers will be removed.

* Removes unused locals

* Removes whitespace

* Propagates `EndEnchantment`

* Removes concentration requirement from description

The caster can move and engage in combat while under the effects of umbrella, so this updates the description to reflect that.