Persists command-driven safety changes to game data
This commits addresses an issue where the use `safetyon` and `safetyoff` display unintended behavior, like turning your safety on or off when equipping or unequipping a item.

Currently, when a player uses these commands they're player flags are updated but their config is not. You can see these by using these commands and looking at your Preferences. Furthermore, there are reproducible steps where unequipping an item can drop your safety and that's not good.

This makes sure that once the player's safety is set, it's persisted to their player and config data.