Added items and updated loot tables
Added items to add variety to monster loot drops:
- Stinky-cap mushroom
- corpse mushroom
- golden mushroom
- bone dust
- blackened bone dust
- avar feather
- troll blood
- mystery meat
- milk
- ancient sword/sword of pure ice*

* this weapon has variable properties, intended to produce a variety of weapon variants as for-flavor monster drops without clogging up the class pool with dozens of new weapons

Changed the following item from non-stackable to stackable to use in loot tables, because it currently has zero use in game, and just because it just makes sense:
- egg shell

Updated loot tables for the following monsters to reduce item clutter in rooms, which would lead to eventual blinking/disappearing objects:
- battered skeleton
- tusked skeleton
- daemon skeleton
- avar warrior
- troll
- narthyl worm

Updated loot tables for the following monsters, mostly just for flavor:
- baby spider
- spider
- giant rat/snow rat (shared table)

Updated the following loot tables because they were lousy for being bosses:
- queen spider

Updated the following loot tables for being over the top compared to the difficulty level of the monster:
- frogman

Added loot tables to the following monsters, which had none until now:
- ice creature
- cow
- guard cow

Also set the unused cowX image resource to cows to give them a corpse, but I'll admit it's pretty low-quality in its current state.  (it's just the normal cow image, facing forward.  May be better off as blank.bgf, but a corpse is needed so the trestype class doesn't freak out.)