Additional commands for rented room management
Currently, room rental is a bit opaque in terms of commands the innkeeper accepts as well as time remaining.  Additionally, control over key copies is pretty much nonexistent, and once stolen, your room is compromised.  This PR enhances these aspects of room management:

- Room Help:  Say this to any innkeeper to get a list of commands related to renting rooms, including prices, maximum rentable days, etc.

- Change Lock:  Say this to an innkeeper you have a room with and they'll destroy any copies of your key that are currently in circulation.

- Rent:  Say this to an innkeeper you have a room with and they'll tell you how much time is left on your room, including a conversion to real life days.

Increased security for copied keys:  Key copies, which are not intended to be dropped on death, shattered, swapped, vaulted, etc can still be dropped accidentally.  This PR changes this behavior so that only the room owner can drop or trade copied keys to other players freely.  Any player who receives a copy of someone else's room key can only offer it back to the room's owner, or give it to an innkeeper to destroy it.