Hackish code for torches in mend fixed
The mend spell has a special exception in its logic to exclude the Torch class which is...gross.  I removed it and added an override for CanMend in the Torch class, because mend calls CanMend to determine if an item can be mended or not, not ReqRepair.

Note that this is a nitpick fix that doesn't affect the actual gameplay in any way.