Removed durability damage from shirts, pants, and skirts (#544)
As cosmetic items, it's a bit silly that shirts, pants, and skirts take durability damage and eventually break with use.  This leads to people preferring to just hunt naked, or to not bother with these items at all due to the inconvenience of having to pick up new ones regularly.

Removing the durability aspect encourages people to dress up and look their best wherever they go--something they can already do with another cosmetic item:  masks.

This changes the following:
- Shirts (shirt, royalshirt, tanktop) and pants (pantsa, pantsb, pantsc, pantsd, skirt, and shortskirt) to no longer suffer damage and a reduction in hits when the wearer is damaged.
- No longer displays their condition in descriptions.
- No longer mendable, since that doesn't matter anymore.