Rentable room update 1: Consolidation of duplicate code found in all rent-enabled innkeepers
Each innkeeper that rents rooms has the same large chunks of copy-pasted logic for parsing commands and handling offered items, and that's sloppy.  This PR consolidates all of that logic into the rented-room maintenance object and now only requires that the innkeeper pass in the appropriate values to the appropriate function calls.

Additionally, Moxal's room rental speech lines were exact copy pastes of Morrigan's, so I changed them to more in character for him.