Adds node attack timer checks
This commit adds two timer checks to the node attack event.

The first check is made with each new game day that passes. Because the randomness of the timer can be greater than the length of an in-game day, I'm almost certain this logic was adding additional `StartAttackTimers`. We confirmed many of these timers exist on live.

The other timer check occurs when an actual node attack is triggered. In that case, any existing `ptAttack` timer is deleted before being the property is reused for sending player notifications.

Ideally, once deployed, this change should stop new and unwanted StartAttackTimers from being created, allow the current set of StartAttackTimers to expire, and settle the node attack event down to its intended cadence.