Fix node attacks by checking for existing `ptAttack` timers (#552)
* Readability improvements

* Revert "Readability improvements"

This reverts commit 5b13cfb0c61b39b3bb04303cd51b2290adb2850f.

* Adds node attack timer checks

This commit adds two timer checks to the node attack event.

The first check is made with each new game day that passes. Because the randomness of the timer can be greater than the length of an in-game day, I'm almost certain this logic was adding additional `StartAttackTimers`. We confirmed many of these timers exist on live.

The other timer check occurs when an actual node attack is triggered. In that case, any existing `ptAttack` timer is deleted before being the property is reused for sending player notifications.

Ideally, once deployed, this change should stop new and unwanted StartAttackTimers from being created, allow the current set of StartAttackTimers to expire, and settle the node attack event down to its intended cadence.

* Adds check to allow existing timers to die off

This commit adds a check so that servers can disable the node attack event and allow any existing "StartAttackTimer" timers to expire without being replaced by another timer. Without this, currently running servers will maintain a constant number of bad timers, of which there might be many.