Shaman Update
Currently, avar shamans are just an annoying pest that only serves to make the zones they inhabit a pain to do anything in.  They're nonthreatening as an attacker, but they'll shatter equipment which...frankly sucks and nobody likes them because of it.

This PR removes shatter from them, and replaces it with an overall increase in toughness as well as different spell lists based on faction.  Overview:

Peet-seeeep (promagic):
- Increased mana pool
- Increased cast frequency
- Increased mana regen time
- Can cast fireball, lightning bolt, heat, and earthquake

Bone faction:
- Slightly increased mana pool
- Slightly increased cast frequency
- Can cast hold, apparition, darkness, and sandstorm

Kyip-kyip-kreeet (antimagic):
- Reduced mana regen time
- Increased speed (tougher combatant)
- Can cast dement, forget, dazzle, and antimagic aura

All:
- Increased level to 130, up from 100
- Known spells are stored in a list which can be edited by creative DMs