Adds secure key copies to rentable rooms
Silver room keys (key copies) are nice because you can share your room with friends, alts, etc, however the risk of them losing it, either on purpose or accident, exists.

This PR adds an alternate type of key copy option, the bronze room key, which the owner of the room can trade freely, but anyone else can only give it back to the owner or to an innkeeper for destruction.