Rentable room invalid mail fix
This adds a check for the expiry warning resource (now intended to be a whisper and thus not formatted for user mails) to CleanBadMail in user.kod.  It should allow CleanBadMail to effectively remove all of the invalid innkeeper expiry warning mails while leaving any others untouched.

After being run once and saving, this change can be safely reverted as it will no longer be relevant.

Send o 0 CleanAllBadMail should allow it to run on every single user in one go.

Testing:  I tested this on a character with an invalid expiry mail plus a mail from another user, as well as a mail from the guardian angel--all unread.  It deleted the expiry mail while leaving the others untouched.