Revert skill improvement penalty in Origin of the Kriipa (#581)
Long ago, the Origin of the Kriipa map had single-file choke points that a player could stand in, hold down attack, and just afk-kill enemies as they lined up.  To discourage this, a skill improvement penalty was added.

The map has since been changed to remove these choke points, so this penalty is no longer necessary.