Fix error messages at startup. (#591)
Users remain technically logged in when the server exits.  When it restarts,
they're logged out, causing them to leave their room and break any
trance.  This sends a message to other objects in the room; however,
at this point in server initialization, the room hash table hasn't yet
been created, causing error messages.

The hash table initialization was moved earlier so that the tables exist
while this logging out process is happening.

Co-authored-by: Meridian59 <meridian59.email@gmail.com>