Fix for room key copies not being garbage collected after being deleted
As long as the master (gold) room key that a key copy is created from exists, that key copy cannot be garbage collected after being deleted (given to an innkeeper for destruction) since it remains in the master key's plCopies list.

This should solve that problem.

Note:  Destruction of the gold room key via room expiration, as well as use of the "change lock" command are unaffected by this issue.