Colhorr weapon att fix
The equations for Colhorr weapons are incorrect. The lack of a +1 on the proficiency value means proficiency maxes out at 99, which leads to a truncation that cuts off a point of damage. This is why Colhorr weapons never 'felt like they did anything' - in many cases they quite literally did not.

Move the +1s

Let's move the +1s so they don't skip the big penalty if you have 0 proficiency.