Give The Aerie Guest House a ROOM_HOMETOWN flag
This may be an extremely obscure thing to notice, but there is custom text to describe a citizen of Ko'catan in that player's bio.

player_citizen_kocatan = " has lived in Ko'catan"

However, the Aerie Guest House is missing a ROOM_HOMETOWN flag, so anyone who actually becomes a citizen of Ko'catan will lose that hometown if they ever cast rescue, logoff pen, or get sent home by an admin. You cannot currently choose Ko'catan as a hometown, but it seems like a natural thing that someone might one day add.

ROOM_HOMETOWN flags have no extra functionality, and are found in many non-chooseable-hometown rooms, like OOG or Brahma & Falcon, so it seems like sometimes the flag was placed haphazardly...