Fix for key copies not disposing correctly.
When given to an innkeeper, room key copies are deleted, but not garbage disposed correctly because a reference to the key remains on the master (gold) room key.  This provides a simple fix for that.

The following tests were made:

- Key copy given to innkeeper to destroy it.
- Key copies bulk-destroyed via the change lock command.
- All keys destroyed upon room expiration.

In every test, a save was performed immediately afterward and the copies were successfully garbage collected.