Officially set go bad time to -1 (#666)
Set the Go Bad Time for potions to -1. Formerly, Jala potions never went bad because they are 'labeled,' and the game does not start go bad timers on labeled spell items upon creation. However, the go bad timer could be potentially be restarted if the potion was sold to an NPC and repurchased. Not that such a thing would matter, because the time calculation was also massively off, resulting in go bad times that would have been over a year in length. Jala potions have not had this mechanic for over 20 years at this point, so I figured it's best just to officially set their go bad time to -1. I also don't think it's a good idea to add such a massive number of potential timers to the game, as every single jala potion would have its own go bad timer of a week... while people are building, that could easily mean hundreds or thousands of timers just for a mechanic that wouldn't really add anything.

Co-authored-by: GarAdminTools <meridian59gar4@gmail.com>