Move HealthTime bound to before Restorate calculation
Restorate has much or all of its benefit eliminated by a bound that comes after it. In Restorate's code, it specifically allows a reduction to 650 ms with enough spellpower, but this is immediately eliminated by a min 1000 bound in CalculateHealthTime's code. Below CalculateHealthTime, in CalculateManaTime, we can see that the bound for iTime comes before Rejuvenate and Mana Focus. Based on that, I believe this is a bug, and it confirms my suspicions that Restorate has not actually functioned for the last 22 years. For players to notice an effect from Restorate, they would have had to keep vigor between a certain range (60 to 100 or so), and then Restorate would have kept their HP regen at the same rate as if they had 100+ vigor. However, players don't behave like that (they keep their vigor high while building), so Restorate essentially did not function.