Update Solomon's (Cor Noth grocer) sell list (#690)
Originally, Solomon was supposed to accept inky caps, chunks of nerudite, and orc pit boss heads as gifts to elevate his mood.  That doesn't work and hasn't for a long time.  There is also little to nothing affecting his mood actually does, even if the player knows about how it works.

This removes that dead code and adds the orc pit boss head and chunk of nerudite to his willing-to-buy list.  Inky caps are already covered by IsObjectSundry.