feat: add +1 rarity count for spellcaster items as defined by waspellt.kod.  This weapon attribute works with waspell.kod.
waspell.kod is where a spell is assigned to an item as an item attribute, and waspelt.kod is where the description is fabricated.  It's possible to have them occur independently so I added HasRarity() = TRUE to both.

This means that if the weapon has a spell assigned to it and it shows up in the description, the rarity grade would be +2 (+1 for each IA).  This is ok, because these items are by definition very rare.