fix: undo propagate in ring IsIdentified override.
Thinking about potions a little more, it makes no sense to propagate IsIdentified on potions since if it is labeled, then it will return TRUE.  So there's no need to run the other checks on item attributes, etc.  We can just return FALSE if the label doesn't match the real name.