feature: add Blakod support for item rarity grades (#699)
* fix: remove "%" in buttons in char.rc

* Revert "fix: remove "%" in buttons in char.rc"

This reverts commit 7ae10f71b18d5f979165c4d4c375dbac434eed5d.

* feat: item rarity grade logic for items in Blakod (excluding rings, potions, and assignment of individual rarities to item attributes and item classes)

* chore: minor tweak to comment in reveal.kod

* fix: made code logic less fragile in ToCliObject.

Removed unused local variable assignment.

Renamed HasRarity to GetRarityCountModifier and changed the return value from boolean to an int.

Adjusted docstrings with periods and spacing.

Changed return value of GetRarityCountModifier to a bound call.

* chore: change lowercase bound call to Bound (uppercase)

* feat: make unidentified rings appear so using an overridden IsIdentified applied to ring.kod

For rings I used propagate for IsIdentified in case a ring has a special item attribute added to it, so that the normal IsIdentified checks would occur where it walks through the list if itematts and checks for any unidentified ones

* feat: make potions are appear unidentified if they are so.  Used existing @IsLabelled function as an override of IsIdentified in the potion class.  It is extremely unlikely a potion will have an item attribute and even less that it will be unidentified, but using propagate instead of return so normal checks apply.

Maybe someone will make a potion that sticks with you when you die, for example.

* fix: undo propagate in ring IsIdentified override.

Thinking about potions a little more, it makes no sense to propagate IsIdentified on potions since if it is labeled, then it will return TRUE.  So there's no need to run the other checks on item attributes, etc.  We can just return FALSE if the label doesn't match the real name.

* fix: moved IsIdentified message override for potions, wands, and scrolls to parent class spelitem.kod.

This means all of those will be checked with superclass function IsLabeled to determine if the item should be assigned unidentified or not.

* chore: refactor IsIdentified to be more efficient based on feedback

* chore: re-add deleted empty line from potion.kod

* chore: refactor GetRarityCountModifier to simply return 0 by default.  Item Attribute child classes can use this message to affect the rarity count of items, but this default message simply returns 0.

* chore: renamed GetMagicRarityGradeCount to just GetRarityGradeCount since we don't want to cause confusion by having a rarity grade (magic) in the name of a function that might return things other than magic