New Room IsLegalMove function, Guildhall wallhack protection
Added an IsLegalMove function to rooms, which generally returns TRUE. For guildhalls, it bounces back any player who attempts to cross from the foyer into the main guildhall outside the proper time window. In about a decade of live with this on OS, it's been essentially foolproof. A few other functions will be needed to prevent attacks and object interaction from the foyer to the main hall, just to be certain.

IsLegalMove can also be used to protect other areas where wallhack offers an advantage, such as grabbing the yeti node without killing the yeti.