Link attacks, opening the door, flipping lever, and taking from chest to foyer status
If players are legally inside the hall, they can now open the door, take from the chest, or flip the lever properly.

Players who are not legally inside the hall - aka in the foyer - cannot attack those inside. Exceptions will need to be made for halls where this is possible, such as Sewer Hideout, Konima's Abandoned Dwelling, Abandoned Warehouse, and Bookmaker's.