Fix for DM-added attributes not being revealed
This PR causes item attributes added through the "dm create itematt <attribute>" command to always be revealed.  Previously, an admin using this command would have to use the reveal spell, which is problematic if one of the added attributes was shroud.

If the old behavior of adding an unrevealed attribute is preferred, the command "dm create unrevealed itematt <attribute>" has been added.