Create a Frogman treasure table
The treasure of Frogmen was the same as the Stone Troll, likely because the Frogmen were not actually placed in the world until recently. This creates a treasure table based on some notes in the Ancient monster loot table. It's generally berries and mushrooms, with a tiny sprinkling of DAFs and useful equipment, to match Skeletons (the opposite karma path to this one) dropping BDS now and again. The diff seed is also similar to Skeletons, but with a slightly higher att chance, since Frogmen are sentient beings who might seek out and wield special weapons.