fix: make reveal and identify work on shroud
when the spell shroud is cast, the IA_SHROUD attribute is added to the item but in an initially revealed / identified form

However when admins add the IA_SHROUD item attribute to an item, it will be unidentified.

This change makes shroud revealable/unidentifiable since it should be by default so, but if not, it is impossible to reveal the item.