Fix for DM-added attributes not being revealed (#749)
* Fix for DM-added attributes not being revealed

This PR causes item attributes added through the "dm create itematt <attribute>" command to always be revealed.  Previously, an admin using this command would have to use the reveal spell, which is problematic if one of the added attributes was shroud.

If the old behavior of adding an unrevealed attribute is preferred, the command "dm create unrevealed itematt <attribute>" has been added.

* Minor adjustments

- Added a boolean variable assignment for revealing attributes closer to the rest of the speech comparison logic.
- Added proper explanation of the arcane techno-sorcery that is item attribute compound values in relation to how "revealed" vs "unrevealed" information is stored in them.

* Code adjustments based on feedback

- Adjusted the dialog parsing to do fewer checks when checking for "create unrevealed itematt."
- Adjusted the logic flow so the check for SAY_DM can be easily extended to encompass the other DM commands in a future PR.