fix: remove GetRarityCountModifier +1 from waspellt.kod
This was causing problems because holy weapon and unholy weapon were showing as rare.

Now that I know more about this item attribute, I know it always has a timer and appears to be different types of enchantments for weapon magic attack types (fire, shock, etc).

In any case, it should always have a timer, and since it does then we do not want to count it towards rarity.