fix: revert timer and IA_MISDIRECTION logic from GetRarityGradeCount
My original hope was to have a catchall for item attributes where we would exclude item attributes that did not have a timer.

The complicating factor is I wanted to do this for item attributes that do have timers in their class but perhaps were changed by an admin adjusting the value to NIL.

I think trying to capture these one-off changes in the "normal" pathway is too difficult and convaluted.

Item attributes have a message called HasTimer which can do this for the class, and also we can simply use GetRarityCountModifier to determine what amount of rarity to add.

Finally, and perhaps most importantly, I noticed that some item attributes have a value that is NOT a timer at position 2, so checking position 2 will cause consistency issues.  I see no way to do what I originally intended so I am reverting this.