fix: remove GetRarityCountModifier from WeapAttSpellType (#765)
* fix: ignore item attributes that have a timer for rarity calculation

* fix: let Artifice count towards rarity grade by not excluding that attribute even though it has a timer.

This is the only exception since artifice is intended to misdirect users into thinking items are more rare than they are.
All other item attributes will not count towards the rarity grade count if they have a timer.

* chore: fix whitespace

* fix: remove GetRarityCountModifier +1 from waspellt.kod

This was causing problems because holy weapon and unholy weapon were showing as rare.

Now that I know more about this item attribute, I know it always has a timer and appears to be different types of enchantments for weapon magic attack types (fire, shock, etc).

In any case, it should always have a timer, and since it does then we do not want to count it towards rarity.

* fix: revert timer and IA_MISDIRECTION logic from GetRarityGradeCount

My original hope was to have a catchall for item attributes where we would exclude item attributes that did not have a timer.

The complicating factor is I wanted to do this for item attributes that do have timers in their class but perhaps were changed by an admin adjusting the value to NIL.

I think trying to capture these one-off changes in the "normal" pathway is too difficult and convaluted.

Item attributes have a message called HasTimer which can do this for the class, and also we can simply use GetRarityCountModifier to determine what amount of rarity to add.

Finally, and perhaps most importantly, I noticed that some item attributes have a value that is NOT a timer at position 2, so checking position 2 will cause consistency issues.  I see no way to do what I originally intended so I am reverting this.