Eliminate wasted defense improvement rolls to fix Stuck Parry problem
There is code when a player kills a monster, it randomly sends an improve chance to block, dodge, or parry. This still happens when the skills are maxed, so lots of improvement chances are wasted. This leads to a situation with Parry that is not obvious. It is why players get Stuck Parry. When a player has other schools, their secondary improvement chance on Parry drops to minimum, as low as 1 in 100. Their primary improvement chance is between 20 and 30 based on Intellect, with up to +15 from Jonas faction. So players must beat a very tough roll often 1 in 100. Then 20-45 in 100. Then we see the situation here, where 70 of 100 improvement chances are going to Block and Dodge even when are they are maxed, so it is 1 in 100, 20-45 in 100, 30 in 100 all must be beat to improve. This combined with only getting 1 attempt per monster kill leads to Parry becoming 'Stuck' and the slowest skill to improve in the game under certain situations.