Updated mollusk asset
This commit cleans up alpha channel bleed in frames of the mollusk artwork and recompiles the entire `bgf` to include the underground state. Previously, this was edited using the third-party BGF editor. This update uses `makebgf` and an updated bbg file. It updates the kod as well, to be in sync with the bgf groups.