Sort objects before rendering for consistent world object ordering (#922)
As we receive objects in different orders from the BSP walk, when we
come to render they may also be sorted by z-when at an identical
location. This causes visual issue with items being sorted differently
depending on the players position in a room - we now ensure objects are
sorted by their ids to create a consistent view.