Performs "already learned" check on skill asks of NPCs (#935)
This PR addresses one of the issues brought up in #491 in which NPCs
responses to skills can be confusing. As it stands, when an NPC is asked
about a skill (e.g. a player curious about their progress) they first
check to see if a players current skills meet or exceed the necessary
level. For example, a player asking for a level 1 spell when they're
already working level 2. The code returns there and responds that the
player is ready. However, it can be the case that a player already has
the skill they're asking about and _that_ check isn't reached. This can
be confusing and it seems to me the response for a skill already learned
should always be _the skill has already been learned_.

So, this PR does a quick loop through spells and skills and triggers the
more appropriate response before continuing into the "can learn" logic,
which I left intact so as not to introduce any new issues.