Fix heat scrolls by updating of target retrieval logic (#932)
In the game currently are two harmful scrolls: heat and earthquake.
Earthquake is disabled and heat is enabled, but fails to cast. The issue
as I understand it, is that as a `SpellItem`, scrolls expect a target.
This seems to work fine for most scrolls, but _harmful_ scrolls/spells
like heat and EQ run through additional logic that fails the spell
because the target is the player themselves.

This PR adds a small bit of logic to spell item casts that says, ignore
any "apply_on" target if this is a scroll. It does makes a bold
assumption that no future scroll will be a targeted spell, which makes
me wonder if I should be checking for a room enchantment spell rather
than a scroll. I'm curious to hear thoughts on that as I'm leaning
towards enchantment.

### The issue
When used, the scroll fails and says:
> There are no proper targets for heat.


https://github.com/user-attachments/assets/bc73ceea-be27-4024-ae6a-25bdce37d1db

### The fix in action


https://github.com/user-attachments/assets/41bf2a6e-c6bd-41c5-85b6-dee700ad2281

Note: I don't know what happened with the commit history of #909, so I
put this PR up in its place.

Closes #728