Fixes erroring faction service timer deletion (#944)
When a player joins a faction, `FactionServiceTimer()` creates a timer
callback to itself to determine faction status and if good, create
another timer to do the same in the future. The problem seen in the
logs, is that when control returns to this method, the reference to the
timer remains but the timer has since been deleted. The result is a
`DeleteTimer()` error.

I'm unable to pinpoint a point in the code where an inadvertent deletion
of this timer exists and am kind of assuming that when the timer expires
and the call to `FactionServiceTimer()` is made, the timer has been
cleaned up and the reference via `ptFactionTimer` remains.

To fix this, I've removed any timer deletion code from
`FactionServiceTimer()` and have given that responsibility to
`FactionServiceRoutine()`.

### Testing
Testing this is made much easier by tuning `FACTION_RESIGN_TIME` and
`FACTION_UPDATE_TIME` way down to create rapid faction service cycles.